fix: provide a no results search query proxy
The EngineProcessors.createEngineProcessors method now requires a new
SearchClientsProxy. While it's not actually used yet, it may be used in
the future, although it won't serve any purpose in ZPT. Therefore, it's
reasonable to implement it with a static implementation that always
provides 0 results.

Please note that this implementation cannot easily be maintained. New
additions to the SearchClientsProxy interface will demand new
implementations in the NoResultsSearchClientsProxy. We should
investigate whether we can find a different place for the proxy to be
passed to the processors.