test: fix test relying availability of implicit test bean
The test setup was relying on the presence of an implicit bean representing the test. Which is not exposed in the `applicationContext` by default. The fix moves all worker setups into dedicated configuration classes to resolve this.

(cherry picked from commit 343db4f0826a9cb56c4a3470329704c9b430f229)